:sparkles: Feat: 파일 업로드 시 용량 필드 초기화
- File 엔티티, ResponseDto capacity 타입 int → long 변경
- FileUploadService에서 파일 용량 저장 로직 추가

+++ complaint-orchestrator/src/main/java/edu/dongguk/complaint/orchestrator/dto/response/FileResponseDto.java
@@ -16,7 +16,7 @@ public record FileResponseDto(
         String name,
 
         @NotNull
-        int capacity,
+        long capacity,
 
         int complaintCount,
